pyproject.toml
[tool.ruff]
# Exclude a variety of commonly ignored directories.
extend-exclude = [
    ".direnv",
    ".git",
    ".ruff_cache",
    "node_modules",
    "migrations",
]
line-length = 120
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
fixable = [
    "ALL",
]
unfixable = []
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.setuptools]
py-modules = []

[project]
name = "taylor-learns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Django>=5.2.8",
    "boto3",
    "celery",
    "celery-progress",
    "celery[redis]",
    "channels",
    "channels_redis",
    "django-allauth[socialaccount,mfa,headless]>=65.13.0",
    "django-anymail[mailgun]", # email services
    "django-browser-reload",
    "django-celery-beat",
    "django-cors-headers",
    "django-environ",
    "django-health-check",
    "django-hijack",
    "django-htmx",
    "django-storages",
    "django-taggit>=6.1.0",
    "django-template-partials",
    "django-vite>=3.1.0",
    "django-waffle",
    "djangorestframework",
    "djangorestframework-api-key",
    "docutils",
    "drf-spectacular",
    "httpx>=0.28.1",
    "litellm<1.77.2",
    "markdown",
    "nh3>=0.2.21",
    "openai",
    "psycopg2-binary",
    "pydantic-ai>=1.0",
    "redis",
    "requests",
    "sentry-sdk[django]>=2.33.2",
    "wagtail",
    "whitenoise[brotli]",
]

[dependency-groups]
dev = [
    "channels[daphne]",
    "django-debug-toolbar",
    "pre-commit",
    "ruff>=0.13.1",
    "pegasus-cli>=0.8",
]
prod = [
    "gunicorn",
    "uvicorn[standard]",
]
